+++ application/config/config.php
@@ -255,14 +255,14 @@ $config['log_date_format'] = 'Y-m-d H:i:s';
 
 /*
 |--------------------------------------------------------------------------
-| Error Templates Directory Path
+| Error Views Directory Path
 |--------------------------------------------------------------------------
 |
 | Leave this BLANK unless you would like to set something other than the default
 | application/views/errors/ folder.  Use a full server path with trailing slash.
 |
 */
-$config['error_templates_path'] = '';
+$config['error_views_path'] = '';
 
 /*
 |--------------------------------------------------------------------------